refactor (user) : 서비스 인터페이스와 구현체를 분리하여 도메인 결합도 감소 및 유연성 향상
