+++ src/main/java/com/backend/immilog/user/application/UserReportService.java
@@ -2,8 +2,8 @@ package com.backend.immilog.user.application;
 
 import com.backend.immilog.global.application.RedisDistributedLock;
 import com.backend.immilog.global.exception.CustomException;
-import com.backend.immilog.user.infrastructure.ReportRepository;
-import com.backend.immilog.user.infrastructure.UserRepository;
+import com.backend.immilog.user.model.interfaces.repositories.ReportRepository;
+import com.backend.immilog.user.model.interfaces.repositories.UserRepository;
 import com.backend.immilog.user.model.entities.Report;
 import com.backend.immilog.user.model.entities.User;
 import com.backend.immilog.user.presentation.request.UserReportRequest;
